Added newsfragment, some more slight refactoring
