+++ newsfragments/1259.misc.rst
@@ -1,0 +1,1 @@
+Removed Pypika from SQL query generation
