+++ parsec/backend/postgresql/vlob_queries/__init__.py
@@ -1,6 +1,10 @@
 # Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
 
-from parsec.backend.postgresql.vlob_queries.write import query_update, query_vlob_updated
+from parsec.backend.postgresql.vlob_queries.write import (
+    query_update,
+    query_vlob_updated,
+    query_create,
+)
 from parsec.backend.postgresql.vlob_queries.maintenance import (
     query_maintenance_save_reencryption_batch,
     query_maintenance_get_reencryption_batch,
@@ -10,8 +14,6 @@ from parsec.backend.postgresql.vlob_queries.read import (
     query_poll_changes,
     query_list_versions,
 )
-from parsec.backend.postgresql.vlob_queries.utils import query_check_realm_access
-from parsec.backend.postgresql.vlob_queries.create import query_create
 
 __all__ = (
     "query_update",
@@ -21,6 +23,5 @@ __all__ = (
     "query_read",
     "query_poll_changes",
     "query_list_versions",
-    "query_check_realm_access",
     "query_create",
 )
