+++ parsec/backend/postgresql/vlob_queries/write.py
@@ -1,8 +1,8 @@
 # Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
 
 import pendulum
-from triopg import UniqueViolationError
 from uuid import UUID
+from triopg import UniqueViolationError
 
 from parsec.api.protocol import DeviceID, OrganizationID
 from parsec.backend.postgresql.utils import (
@@ -13,9 +13,14 @@ from parsec.backend.postgresql.utils import (
     q_realm_internal_id,
     q_vlob_encryption_revision_internal_id,
 )
-from parsec.backend.vlob import VlobVersionError, VlobTimestampError, VlobNotFoundError
+from parsec.backend.vlob import (
+    VlobVersionError,
+    VlobTimestampError,
+    VlobNotFoundError,
+    VlobAlreadyExistsError,
+)
 from parsec.backend.postgresql.handler import send_signal
-from parsec.backend.postgresql.vlob_queries.utils import query_get_realm_id_from_vlob_id
+from parsec.backend.postgresql.vlob_queries.utils import _get_realm_id_from_vlob_id
 from parsec.backend.backend_events import BackendEvent
 
 
@@ -41,25 +46,26 @@ RETURNING index
 async def query_vlob_updated(
     conn, vlob_atom_internal_id, organization_id, author, realm_id, src_id, src_version=1
 ):
+    async with conn.transaction():
 
-    index = await conn.fetchval(
-        *q_vlob_updated(
-            organization_id=organization_id,
-            realm_id=realm_id,
-            vlob_atom_internal_id=vlob_atom_internal_id,
+        index = await conn.fetchval(
+            *q_vlob_updated(
+                organization_id=organization_id,
+                realm_id=realm_id,
+                vlob_atom_internal_id=vlob_atom_internal_id,
+            )
         )
-    )
 
-    await send_signal(
-        conn,
-        BackendEvent.REALM_VLOBS_UPDATED,
-        organization_id=organization_id,
-        author=author,
-        realm_id=realm_id,
-        checkpoint=index,
-        src_id=src_id,
-        src_version=src_version,
-    )
+        await send_signal(
+            conn,
+            BackendEvent.REALM_VLOBS_UPDATED,
+            organization_id=organization_id,
+            author=author,
+            realm_id=realm_id,
+            checkpoint=index,
+            src_id=src_id,
+            src_version=src_version,
+        )
 
 
 _q_get_vlob_version = Q(
@@ -121,7 +127,7 @@ async def query_update(
     from parsec.backend.postgresql.vlob import _check_realm_and_write_access
 
     async with conn.transaction():
-        realm_id = await query_get_realm_id_from_vlob_id(conn, organization_id, vlob_id)
+        realm_id = await _get_realm_id_from_vlob_id(conn, organization_id, vlob_id)
         await _check_realm_and_write_access(
             conn, organization_id, author, realm_id, encryption_revision
         )
@@ -159,4 +165,78 @@ async def query_update(
 
         await query_vlob_updated(
             conn, vlob_atom_internal_id, organization_id, author, realm_id, vlob_id, version
+        )
+
+
+_q_create = Q(
+    f"""
+INSERT INTO vlob_atom (
+    organization,
+    vlob_encryption_revision,
+    vlob_id,
+    version,
+    blob,
+    size,
+    author,
+    created_on
+)
+SELECT
+    { q_organization_internal_id("$organization_id") },
+    {
+        q_vlob_encryption_revision_internal_id(
+            organization_id="$organization_id",
+            realm_id="$realm_id",
+            encryption_revision="$encryption_revision"
+        )
+    },
+    $vlob_id,
+    1,
+    $blob,
+    $blob_len,
+    { q_device_internal_id(organization_id="$organization_id", device_id="$author") },
+    $timestamp
+RETURNING _id
+"""
+)
+
+
+@query()
+async def query_create(
+    conn,
+    organization_id: OrganizationID,
+    author: DeviceID,
+    realm_id: UUID,
+    encryption_revision: int,
+    vlob_id: UUID,
+    timestamp: pendulum.Pendulum,
+    blob: bytes,
+) -> None:
+    from parsec.backend.postgresql.vlob import _check_realm_and_write_access
+
+    async with conn.transaction():
+
+        await _check_realm_and_write_access(
+            conn, organization_id, author, realm_id, encryption_revision
+        )
+
+        # Actually create the vlob
+        try:
+            vlob_atom_internal_id = await conn.fetchval(
+                *_q_create(
+                    organization_id=organization_id,
+                    author=author,
+                    realm_id=realm_id,
+                    encryption_revision=encryption_revision,
+                    vlob_id=vlob_id,
+                    blob=blob,
+                    blob_len=len(blob),
+                    timestamp=timestamp,
+                )
+            )
+
+        except UniqueViolationError:
+            raise VlobAlreadyExistsError()
+
+        await query_vlob_updated(
+            conn, vlob_atom_internal_id, organization_id, author, realm_id, vlob_id
         )
